pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "v6-diagnostics"
version = "1.0.0"
description = "vantage6 node diagnostic tools"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "vantage6-algorithm-tools>=5.0.0a43",
  "vantage6-client>=5.0.0a43",
  "requests",
  "pyjwt",
  "rich",
  "click",
]
license = { text = "Apache License 2.0" }
keywords = ["vantage6", "algorithm", "diagnostics"]
classifiers = [
  "License :: OSI Approved :: Apache Software License",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.13",
]

[tool.hatch.build.targets.wheel]
packages = ["v6_diagnostics"]